+++ R/hugo.R
@@ -200,9 +200,13 @@ new_content = function(path, kind = 'default', open = interactive()) {
 default_kind = function(path) {
   path = normalizePath(path, '/', mustWork = FALSE)
   if (!grepl('/', path)) return('default')
-  atype = gsub('/.*', '.md', path)
-  if (!file.exists(file.path('archetypes', atype))) return('default')
-  gsub('/.*', '', path)
+  ## Assumes path is something like
+  ## /pathToBlogdownDir/content/post/2018-02-24-postslug.Rmd
+  ## (file termination doesn't matter)
+  if(!grepl('/content/', path)) return('default')
+  atype = gsub('.*/', '', dirname(path))
+  if (!file.exists(file.path('archetypes', paste0(atype, '.md')))) return('default')
+  atype
 }
 
 # Hugo cannot convert a single file: https://github.com/gohugoio/hugo/issues/3632