+++ inst/scripts/new_post.R
@@ -35,6 +35,12 @@ local({
         ),
         height = '70px'
       ),
+      shiny::fillRow(
+        shiny::selectInput(inputId = 'kind', label = 'Archetype',
+          choices = c('default', gsub('.md', '', dir('archetypes',
+            pattern = '\\.md$')))),
+        height = '70px'
+      ),
       miniUI::gadgetTitleBar(NULL)
     )),
     server = function(input, output, session) {
@@ -66,7 +72,8 @@ local({
           input$title, author = input$author, ext = input$format,
           categories = input$cat, tags = input$tag,
           file = gsub('[-[:space:]]+', '-', input$file),
-          slug = input$slug, subdir = input$subdir, date = input$date
+          slug = input$slug, subdir = input$subdir, date = input$date,
+          kind = input$kind
         )
         shiny::stopApp()
       })
@@ -74,6 +81,6 @@ local({
         shiny::stopApp()
       })
     },
-    stopOnCancel = FALSE, viewer = shiny::dialogViewer('New Post', height = 500)
+    stopOnCancel = FALSE, viewer = shiny::dialogViewer('New Post', height = 570)
   )
 })